+++ LeezeIt/LeezeIt/Controller/PhoneNumberViewController.swift
@@ -44,16 +44,6 @@ class PhoneNumberViewController:UIViewController{
         present(nextvc, animated: true)
     }
     
-    func fetchUserDefaultData () {
-        let valueOfPhoneNumber = UserDefaults.standard.string(forKey: "userPhoneNumber")
-        if valueOfPhoneNumber != nil {
-            phoneNumberText.text = valueOfPhoneNumber
-        }
-    }
-}
-
-extension PhoneNumberViewController : UITextFieldDelegate {
-    
     func format(with mask: String, phone: String) -> String {
         let numbers = phone.replacingOccurrences(of: "[^0-9]", with: "", options: .regularExpression)
         var result = ""
@@ -72,17 +62,16 @@ extension PhoneNumberViewController : UITextFieldDelegate {
         return result
     }
     
-    func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
-        guard let text = textField.text else { return false }
-        let newString = (text as NSString).replacingCharacters(in: range, with: string)
-        textField.text = format(with: "+X (XXX) XXX-XXXX", phone: newString)
-        dismissKeyBoard()
-        return false
-    }
-  
     //Hides keyboard whenever user click/tap anywhere on the screen
     override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
         view.endEditing(true)
+    }
+    
+    func fetchUserDefaultData () {
+        let valueOfPhoneNumber = UserDefaults.standard.string(forKey: "userPhoneNumber")
+        if valueOfPhoneNumber != nil {
+            phoneNumberText.text = valueOfPhoneNumber
+        }
     }
     
     //hide keyboard after user entered 10 digits of phoneNumber
