+++ LeezeIt/LeezeIt/Controller/SignUpViewController.swift
@@ -7,7 +7,7 @@
 
 import UIKit
 
-class SignUpViewController: UIViewController, UITextFieldDelegate {
+class SignUpViewController: UIViewController {
     
     @IBOutlet weak var firstNameTextField: UITextField!
     @IBOutlet weak var lastNameTextField: UITextField!
@@ -27,31 +27,11 @@ class SignUpViewController: UIViewController, UITextFieldDelegate {
     }
 }
 
+// MARK:    Extension SignUPViewController
+
 extension SignUpViewController {
     
-    func hideKeyboardWhenTappedView() {
-        let tap: UITapGestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(dismissKeyboard))
-        self.view.addGestureRecognizer(tap)
-    }
-    
-    @objc func dismissKeyboard() {
-        self.view.endEditing(true)
-    }
-    
-    func checkEmailFormat(email: String) -> Bool {
-        let emailEntries = "[A-Z0-9a-z._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,64}"
-        let emailFormat = NSPredicate(format: "SELF MATCHES %@", emailEntries)
-        return emailFormat.evaluate(with: email)
-    }
-    
-    //        restrict user to enter numbers in name textFields
-     func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
-        if textField == firstNameTextField || textField == lastNameTextField {
-            return checkTextField(string: string)
-        }
-        return true
-    }
-    
+
     func checkTextField(string: String) -> Bool {
         let allowedCharacters = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
         let allowedCharacterSet = CharacterSet(charactersIn: allowedCharacters)
@@ -82,11 +62,18 @@ extension SignUpViewController {
         [passwordTextField, emailTextField].forEach({ $0.addTarget(self, action: #selector(editingChanged), for: .editingChanged) })
     }
     
-    @objc func editingChanged(_ textField: UITextField) {
-        
-        let validEmail = checkEmailFormat(email: emailTextField.text!)
-        if validEmail == true && passwordTextField.text!.count > 6 {
-            nextButton.isEnabled = true
-        }
+    func hideKeyboardWhenTappedView() {
+        let tap: UITapGestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(dismissKeyboard))
+        self.view.addGestureRecognizer(tap)
+    }
+    
+    @objc func dismissKeyboard() {
+        self.view.endEditing(true)
+    }
+    
+    func checkEmailFormat(email: String) -> Bool {
+        let emailEntries = "[A-Z0-9a-z._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,64}"
+        let emailFormat = NSPredicate(format: "SELF MATCHES %@", emailEntries)
+        return emailFormat.evaluate(with: email)
     }
 }
